9491 - genericized intergration test to handle both section and my queues for message boxes
