+++ shared/src/business/entities/EntityConstants.js
@@ -1233,6 +1233,8 @@ const OTHER_FILER_TYPES = [
 
 const CASE_MESSAGE_DOCUMENT_ATTACHMENT_LIMIT = 5;
 
+const MESSAGE_QUEUE_TYPES = ['my', 'section'];
+
 const TRIAL_SESSION_ELIGIBLE_CASES_BUFFER = 50;
 
 const ADVANCED_SEARCH_TABS = {
@@ -1332,6 +1334,7 @@ module.exports = deepFreeze({
   MAX_FILE_SIZE_MB,
   MAX_SEARCH_CLIENT_RESULTS: 200,
   MAX_SEARCH_RESULTS: 100, // a fraction of MAX_SEARCH_CLIENT_RESULTS
+  MESSAGE_QUEUE_TYPES,
   MINUTE_ENTRIES_MAP,
   NOTICE_OF_CHANGE_CONTACT_INFORMATION_EVENT_CODES,
   NOTICE_OF_CHANGE_CONTACT_INFORMATION_MAP,
